refactor(Table): TableScrollerの内部実装を最適化 (#6474)
